Fixed first issues with database being wrongfully generated. Still need to fix modal interactions

+++ cogs/artwork.py
@@ -16,32 +16,40 @@ class Artwork(commands.Cog):
             channels_dict = json.load(f)
 
         # Get the server ID
-        server_id = str(message.guild.id)
+        if message.guild is not None:
+            server_id = str(message.guild.id)
 
-        if server_id in channels_dict:
-            allowed_channels = channels_dict[server_id]
-            if str(message.channel.id) in allowed_channels and len(message.attachments) > 0:
-                # Create a thread
-                thread = await message.create_thread(name=f"Artwork Discussion of {message.author.name}")
-                await thread.send(f"Look at what {message.author.mention} just made!")
+            if server_id in channels_dict:
+                allowed_channels = channels_dict[server_id]
+                if str(message.channel.id) in allowed_channels and len(message.attachments) > 0:
+                    # Create a thread
+                    thread = await message.create_thread(name=f"Artwork Discussion of {message.author.name}")
+                    await thread.send(f"Look at what {message.author.mention} just made!")
+        else:
+            pass
 
     @commands.Cog.listener("on_message")
     async def check_attachment(self, message):
         # If the file exists, load the data from the file
         with open('channels.json', 'r') as f:
             channels_dict = json.load(f)
-        # Get the server ID
-        server_id = str(message.guild.id)
 
-        if server_id in channels_dict:
-            allowed_channels = channels_dict[server_id]
+        if message.guild is not None:
+            # Get the server ID
+            server_id = str(message.guild.id)
 
-            if str(message.channel.id) in allowed_channels and len(message.attachments) == 0 and not message.author.guild_permissions.administrator:
-                # Delete the message
-                await message.delete()
-                # Send a message to the user discreetly
-                user = message.author
-                await user.send(f"You cannot send messages without attachments in {message.channel.mention}.")
+            if server_id in channels_dict:
+                allowed_channels = channels_dict[server_id]
+
+                if str(message.channel.id) in allowed_channels and len(message.attachments) == 0 and not message.author.guild_permissions.administrator:
+                    # Delete the message
+                    await message.delete()
+                    # Send a message to the user discreetly
+                    user = message.author
+                    await user.send(f"You cannot send messages without attachments in {message.channel.mention}.")
+
+        else:
+            pass
 
 
 async def setup(bot):
